statx: add STATX_FHANDLE_HASH for hashed file handle
As inode numbers are no longer guaranteed to be unique on certain
filesystems, and userspace being pushed towards using file handles to
check whether two files are the same inode, userspace now has to do
something like the following to safely get the file handle of an unknown
file (assuming they need to get some other information from statx(2) and
they want to make sure the file handle references the same file, safe
from races):

  int fd = openat(-EBADF, "/foo/bar/baz", O_PATH | O_CLOEXEC);
  err1 = name_to_handle_at(fd, "", &handle, &unused_mntid, AT_EMPTY_PATH);
  err2 = statx(fd, "", AT_EMPTY_PATH, STATX_MNT_ID_UNIQUE, &statxbuf);
  close(fd);

which is at least 4x more syscalls than just using fstatat(2) or
statx(2) to get the device and inode number. For users that have to open
files already, this is isn't too large of a burden, but for tools that
are scanning directories with statx(2) adding the need to open files is
quite unfortunate.

Making it possible get a file handle with statx(2) would make this
change in behaviour more palletable to userspace, and would properly
signal to authors of new programs that stx_ino is no longer something
that can reasonably be used.

Unfortunately, file handles can be quite large (<=128 bytes at the
moment) and our uAPI allows the maximum size to be increase in the
future. While this means we cannot just shove the handle into struct
statx, userspace only cares about the uniqueness between two inodes, and
so we can just hash the file handle and store the hash.

To avoid wasting too many bits of struct statx, we use a Blake2s-192
hash. Assuming Blake2s is a strong hash, we would expect there to be a
1-in-2^(-63) chance of collision with 2^64 inodes and 1-in-a-million
with 2^86 inodes. This should be more than sufficient for any real
filesystem in the forseeable future.

Suggested-by: Lennart Poettering <mzxreary@0pointer.de>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>