cgroup: add mount option to specify cgroupns of new mount
Being able to specify the cgroupns for a new cgroup mount makes using
cgroup namespaces more ergonomic, especially when configuring mounts for
a container.

Compare

  sfd = fsopen("cgroup2", 0);
  fsconfig(sfd, FSCONFIG_SET_FD, "cgroupns", NULL, nsfd);
  fsconfig(sfd, FSCONFIG_CREATE, NULL, NULL, 0);

versus

  oldns = open("/proc/self/ns/cgroup", O_RDONLY);
  setns(nsfd, CLONE_NEWCGROUP);
  sfd = fsopen("cgroup2", 0);
  setns(oldns, CLONE_NEWCGROUP);
  close(oldns);
  fsconfig(sfd, FSCONFIG_CREATE, NULL, NULL, 0);

In green-thread-based programming languages like Go, the second example
would also require locking the thread to the correct OS thread and some
other considerations to make sure that the superblock doesn't get
instantiated with the wrong cgroup namespace.

To avoid any possible invalidation issues, we do not allow an existing
mount to have its cgroupns reconfigured. If a process needs to do this,
it is fairly trivial to create a new cgroup{,2} mount with the correct
cgroupns and use MOVE_BENEATH to "swap" the mount.

An alternative interface would be to use FSCONFIG_SET_PATH{,_EMPTY} to
allow for passing paths as well as just file descriptors, but there
isn't much infrastructure around that at the moment. If necessary, we
can always add it as a backwards-compatible change. Since namespace
interfaces primarily use file descriptors, let's stick with that
interface for now.

Cc: Christian Brauner <brauner@kernel.org>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>