cgroup: do not change fc->user_ns for FSCONFIG_RECONFIGURE
When reconfiguring an existing mount we do not change the cgroupns of
the mount, so it seems prudent to not update in that case fc->user_ns
(which affects the permission checking).

In practice fc->user_ns will almost always match the sb->s_user_ns of
the cgroupfs, but it seems there might be cases where this would've
allowed a less privileged user to modify the flags of an existing cgroup
mount for a cgroupns we don't have privileges over.

Cc: David Howells <dhowells@redhat.com>
Fixes: 23bf1b6be9c2 ("kernfs, sysfs, cgroup, intel_rdt: Support fs_context")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>